refactoring : redis scan 범위 추가, jpa scan 범위 추가, redis, jpa 레포지토리 분리
