+++ src/main/java/com/moco/moco/config/RedisConfig.java
@@ -5,8 +5,10 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
+import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
 
 @Configuration
+@EnableRedisRepositories(basePackages = "com.moco.moco.redisRepository")
 public class RedisConfig {
 	@Value("${spring.data.redis.host}")
 	private String host;
