+++ src/main/java/com/moco/moco/jpaRepository/BannerRepository.java
@@ -1,9 +1,9 @@
-package com.moco.moco.repository;
+package com.moco.moco.jpaRepository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.moco.moco.domain.Banner;
 
 public interface BannerRepository extends JpaRepository<Banner, Long> {
-	
+
 }
