+++ src/main/java/com/moco/moco/jpaRepository/BookmarkRepository.java
@@ -1,4 +1,4 @@
-package com.moco.moco.repository;
+package com.moco.moco.jpaRepository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
