+++ src/main/java/com/moco/moco/service/user/UserService.java
@@ -15,7 +15,7 @@ import com.moco.moco.dto.UserDto;
 import com.moco.moco.dto.auth.TokenDto;
 import com.moco.moco.exception.CustomAuthenticationException;
 import com.moco.moco.exception.ErrorCode;
-import com.moco.moco.repository.UserRepository;
+import com.moco.moco.jpaRepository.UserRepository;
 import com.moco.moco.service.jwt.JwTokenService;
 import com.moco.moco.service.oauth.OauthService;
